Reorganized folders structure
