+++ __init__.py
@@ -1,10 +1,10 @@
-from .src.agent.base import BaseAgent, LangchainBaseAgent
-from .src.convo.messages import CatMessage, AgentOutput, CatToolMessage, LLMAction
+from .src.agents import BaseAgent, LangchainBaseAgent
+from .src.messages import CatToolMessage, LLMAction
+from cat.agents.base_agent import AgentOutput
 
 __all__ = [
     "BaseAgent",
-    "LangchainBaseAgent",
-    "CatMessage",
+    "LangchainBaseAgent", 
     "AgentOutput",
     "CatToolMessage",
     "LLMAction",
