+++ src/messages.py
@@ -1,26 +1,35 @@
+"""
+Agent Factory - Message Classes
+
+This module contains message and data structure classes used by agents
+for communication and tool calling functionality.
+"""
 
 from typing import Dict, List
 
 from pydantic import BaseModel
-
 from langchain_core.messages import AIMessage, ToolMessage, ToolCall
 
 from cat.agents.base_agent import AgentOutput
 from cat.convo.messages import CatMessage
 
+
 class LLMAction(BaseModel):
+    """Represents an action (tool call) requested by the LLM."""
     id: str
     name: str
     input: Dict
 
+
 class CatToolMessage(CatMessage):
+    """Message representing a tool call and its result in the chat history."""
     action: LLMAction
     result: AgentOutput
 
     def langchainfy(self) -> List[ToolMessage]:
-
-        # Message to rapresent the tool called
-        tool_call =  AIMessage(
+        """Convert to LangChain format for chat history."""
+        # Message to represent the tool called
+        tool_call = AIMessage(
             content=f"Tool Call: {self.action.name}",
             tool_calls=[
                 ToolCall(
